utils: `ImeUtils`: new, move from PrefMainActivity's companion object
